Initial Create
Created the project in Android Studio, with a min SDK, API 24.

Implemented / enabled view binding, for easier access of view components, maintaining coding consistency.

Implemented the necessary Android navigation dependencies, which will be used for navigation around the application.

Structured the project and added packages to separate components of the application, for better maintainability and readability.