feat: Added navigation from dashboard
Added navigation from the dashboard to:
- Tasks
- Simple List
- Events
- Settings

Also added back navigation (navigate up) to go back to the dashboard from the other fragments.

Implemented a top tool bar, so display the name of the fragment, and a back arrow to redirect the user back to the dashboard.

Created a "More" button in the tool bar that displays a menu of all the app's features (no functionality currently) - To be implemented later.