feat: Added bottom nav-bar & fixed UI inconsistencies
Added a bottom-navigation bar to the MainActivity and created a menu for it for the icons / items that will be used for navigation (not currently functional).

Fixed some UI inconsistencies that were noticeable.

Created a SimpleListAdapter which will be implemented later.

Created a fragment for the Login and the GetStarted view.

Login functionality to be implemented later.

Updated: Navigation starts at the Login -> Get Started and then to Dashboard where the rest of the navigation is available.